requirements.txt
sqlalchemy
pgvector
psycopg[binary]
pypdf
typer
phi
phidata
openai
groq
fastapi
uvicorn
duckduckgo-search
streamlit
google-generativeai
